+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 - Use Concurrent.usable_processor_count when it is available ([#2339](https://github.com/getsentry/sentry-ruby/pull/2339))
 
+### Bug Fixes
+
+- Suppress the unnecessary “unsupported options notice” ([#2349](https://github.com/getsentry/sentry-ruby/pull/2349))
+
 ## 5.18.1
 
 ### Bug Fixes